refactor: 메인 페이지 auction component 재사용하도록 함.
